+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/DefaultFilterTableSearchModel.java
@@ -173,7 +173,7 @@ final class DefaultFilterTableSearchModel<C> implements FilterTableSearchModel {
 		Predicate<String> predicate = searchPredicate.get();
 		for (int row = 0; row < tableModel.getRowCount(); row++) {
 			for (int column = 0; column < tableModel.getColumnCount(); column++) {
-				if (predicate.test(tableModel.getStringAt(row, columnModel.getColumn(column).getIdentifier()))) {
+				if (predicate.test(tableModel.getStringAt(row, columnModel.getColumn(column).identifier()))) {
 					searchResults.add(new DefaultRowColumn(row, column));
 				}
 			}