+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTableHeaderRenderer.java
@@ -59,9 +59,9 @@ final class FilterTableHeaderRenderer<R, C> implements TableCellRenderer {
 		if (component instanceof JLabel) {
 			JLabel label = (JLabel) component;
 			FilterTableColumn<C> tableColumn = filterTable.getColumnModel().getColumn(column);
-			ColumnConditionModel<?, ?> filterModel = filterTable.getModel().filterModel().conditionModels().get(tableColumn.getIdentifier());
+			ColumnConditionModel<?, ?> filterModel = filterTable.getModel().filterModel().conditionModels().get(tableColumn.identifier());
 			label.setFont((filterModel != null && filterModel.enabled().get()) ? defaultFont.deriveFont(Font.ITALIC) : defaultFont);
-			label.setIcon(sortArrowIcon(tableColumn.getIdentifier(), label.getFont().getSize() + SORT_ICON_SIZE));
+			label.setIcon(sortArrowIcon(tableColumn.identifier(), label.getFont().getSize() + SORT_ICON_SIZE));
 			label.setIconTextGap(0);
 			if (columnCellRenderer instanceof JLabel) {
 				label.setHorizontalAlignment(((JLabel) columnCellRenderer).getHorizontalAlignment());